First integration test for grpc
To achieve this move the implementation classes in the grpc test
to actual production code.

Always insert a genesis block into the inmemory database.
Moved some more code to our db interface rather than the sql
implementaiton.

Also removed some last compiler error regarding bluethooth
