--- app/src/main/java/nl/tudelft/cs4160/identitychain/connection/GrpcCommunication.kt
@@ -1,6 +1,0 @@
-package nl.tudelft.cs4160.identitychain.connection
-
-class GrpcCommunication() {
-    val C
-
-}
